+++ examples/next/getting-started/src/pages/category/[categorySlug]/[paginationTerm]/[categoryCursor].tsx
@@ -1,6 +1,6 @@
 import { GetStaticPropsContext } from 'next';
 import Page from 'pages/category/[categorySlug]';
-import { getNextStaticProps } from '@wpengine/headless-next';
+import { getNextStaticProps } from '@faustjs/next';
 import { client } from 'client';
 
 export default Page;