+++ examples/next/getting-started/src/pages/posts/index.tsx
@@ -1,4 +1,4 @@
-import { getNextStaticProps } from '@wpengine/headless-next';
+import { getNextStaticProps } from '@faustjs/next';
 import { client, OrderEnum, PostObjectsConnectionOrderbyEnum } from 'client';
 import { Footer, Header, Pagination, Posts } from 'components';
 import { GetStaticPropsContext } from 'next';