+++ examples/react/getting-started/src/client/index.ts
@@ -2,7 +2,7 @@
  * GQLESS: You can safely modify this file and Query Fetcher based on your needs
  */
 import type { IncomingMessage } from 'http';
-import { getClient } from '@wpengine/headless-react';
+import { getClient } from '@faustjs/react';
 import {
   generatedSchema,
   scalarsEnumsHash,